Move the various readers to their own files
